churn/requirements.txt
google-cloud-logging
google-cloud-aiplatform
google-cloud-bigquery
pandas
pyarrow
scikit-learn
kfp
